Fix compile time warnings

+++ src/Game/Render/Canvas.hpp
@@ -1,12 +1,14 @@
 #pragma once
 
+#include <Utils/Math.hpp>
+
 struct Canvas
 {
     static Canvas From(float fov, float distance = 1)
     {
         return Canvas {
             fov,
-            tan(fov / 2) * distance * 2,
+            tanf(fov / 2) * distance * 2,
             distance,
         };
     }
